# docs/backend/nivesh-app-openapi.yaml
openapi: 3.1.0
info:
  title: Nivesh App API
  version: 2.0.0
  description: |
    REST API for the Nivesh.ai application — portfolio management, goal planning,
    AI copilot, broker integration, MFD workspace, and admin controls.

    ## Authentication

    All endpoints (except `/api/auth/*`, `/api/public/*`, and `/api/cas-parser-provider/active`)
    require a session cookie set by `POST /api/auth/google`.

    Admin endpoints additionally require the requesting user to have `is_admin = true`.

    ## Base URL

    - **Production**: `https://niveshcopilot.com`
    - **Staging**: `https://staging.niveshcopilot.com`

  contact:
    name: Nivesh
    url: https://nivesh.com
servers:
  - url: https://niveshcopilot.com
    description: Production
  - url: https://staging.niveshcopilot.com
    description: Staging
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie set by `POST /api/auth/google`.
...
security:
  - SessionCookie: []
paths:
...
  /api/goals/snapshot:
    get:
      tags: [goals]
      operationId: goals_get_snapshot
      summary: Get financial snapshot
      responses:
        "200":
          description: Financial snapshot

    put:
      tags: [goals]
      operationId: goals_upsert_snapshot
      summary: Create or update financial snapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                age:
                  type: integer
                marital_status:
                  type: string
                dependents:
                  type: integer
                monthly_income_rs:
                  type: number
                income_growth_pct:
                  type: number
                monthly_expenses_rs:
                  type: number
                inflation_pct:
                  type: number
                current_corpus_rs:
                  type: number
                total_liabilities_rs:
                  type: number
                risk_profile:
                  type: string
                behavior_score:
                  type: number
      responses:
        "200":
          description: Updated snapshot
  /api/goals:
    get:
      tags: [goals]
      operationId: goals_list
      summary: List all goals with on-track status
      responses:
        "200":
          description: Goals list

    post:
      tags: [goals]
      operationId: goals_create
      summary: Create a new goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [goal_type, goal_name, target_amount_rs, horizon_years, priority]
              properties:
                goal_type:
                  type: string
                  enum: [retirement, education, home, emergency, wealth, other]
                goal_name:
                  type: string
                target_amount_rs:
                  type: number
                horizon_years:
                  type: integer
                priority:
                  type: string
                  enum: [high, medium, low]
                inflation_pct:
                  type: number
                current_corpus_rs:
                  type: number
                monthly_sip_rs:
                  type: number
                manual_allocation:
                  type: object
                  nullable: true
      responses:
        "200":
          description: Created goal
  /api/goals/fund-shortlist/{bucket}:
    get:
      tags: [goals]
      operationId: goals_fund_shortlist
      summary: Top-N funds for a goal allocation bucket
      parameters:
        - name: bucket
          in: path
          required: true
          schema:
            type: string
            example: large_cap
        - name: n
          in: query
          schema:
            type: integer
            default: 15
        - name: min_quality
          in: query
          schema:
            type: number
            default: 55.0
      responses:
        "200":
          description: Fund shortlist
  /api/goals/{goal_id}:
    get:
      tags: [goals]
      operationId: goals_get
      summary: Get goal detail with simulation
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Goal detail

    patch:
      tags: [goals]
      operationId: goals_update
      summary: Update goal fields
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                goal_name:
                  type: string
                target_amount_rs:
                  type: number
                horizon_years:
                  type: integer
                priority:
                  type: string
                monthly_sip_rs:
                  type: number
                current_corpus_rs:
                  type: number
                allocation:
                  type: object
                selected_funds:
                  type: array
                  items:
                    type: string
                manual_fund_override:
                  type: boolean
                status:
                  type: string
      responses:
        "200":
          description: Updated goal

    delete:
      tags: [goals]
      operationId: goals_archive
      summary: Archive (abandon) a goal
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archive result
  /api/goals/{goal_id}/simulate:
    post:
      tags: [goals]
      operationId: goals_simulate
      summary: Run fresh Monte Carlo simulation for a goal
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Simulation result with on_track_pct and scenario outcomes
  /api/goals/{goal_id}/what-if:
    post:
      tags: [goals]
      operationId: goals_what_if
      summary: Preview-only evaluation (does not persist changes)
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                monthly_sip_rs:
                  type: number
                horizon_years:
                  type: integer
                current_corpus_rs:
                  type: number
                allocation_override:
                  type: object
      responses:
        "200":
          description: Preview evaluation
  /api/goals/{goal_id}/copilot:
    post:
      tags: [goals]
      operationId: goals_copilot_ask
      summary: Natural-language what-if via Copilot
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
      responses:
        "200":
          description: Engine output + plain-English narrative
  /api/goals/{goal_id}/copilot/history:
    get:
      tags: [goals]
      operationId: goals_copilot_history
      summary: Load copilot conversation history for a goal
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat history

    delete:
      tags: [goals]
      operationId: goals_copilot_history_clear
      summary: Clear copilot history for a goal
      parameters:
        - name: goal_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result

  # ══════════════════════════════════════════════════════════════════════════
  # Scenarios
  # ══════════════════════════════════════════════════════════════════════════
...
  /api/copilot/ask:
    post:
      tags: [copilot]
      operationId: copilot_ask
      summary: Free-form Q&A with portfolio context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query:
                  type: string
                context:
                  type: object
                  nullable: true
                model:
                  type: string
                  nullable: true
      responses:
        "200":
          description: AI answer
...
  /api/copilot/agents/route:
    post:
      tags: [copilot_agents]
      operationId: copilot_route_intent
      summary: Deterministic intent router — classify message to agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Routed agent ID and spec
...